requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
psycopg2-binary
playwright
beautifulsoup4
lxml
tenacity
python-dateutil
tqdm
